FF:FF的定时任务 延时开启uvc ServerConnoct 更新数据库

+++ controller/src/main/java/com/dcm360/controller/gs/controller/bean/paths_bean/VirtualObstacleBean.java
@@ -640,22 +640,22 @@ public class VirtualObstacleBean implements Cloneable {
                  * x : 118
                  * y : 110
                  */
-                private int x;
-                private int y;
+                private double x;
+                private double y;
 
-                public void setX(int x) {
+                public void setX(double x) {
                     this.x = x;
                 }
 
-                public void setY(int y) {
+                public void setY(double y) {
                     this.y = y;
                 }
 
-                public int getX() {
+                public double getX() {
                     return x;
                 }
 
-                public int getY() {
+                public double getY() {
                     return y;
                 }
             }
